update oauth login

+++ src/main.js
@@ -31,6 +31,59 @@ fitnessHttp.interceptors.response.use(res => {
     // if token not exist, store the token
     if (res.data.token) {
       console.log(`Initilizing token: ${res.data.token}`);
+      window.localStorage.setItem('oauth_token', res.data.token);
+      const Toast = swal.mixin({
+        toast: true,
+        position: 'top-end',
+        showConfirmButton: false,
+        timer: 3000
+      });
+      
+      Toast.fire({
+        type: 'success',
+        title: 'Oauth Signed in successfully'
+      });
+    } else {
+      console.log(`No Response token`);
+    }
+    // setup login status
+    store.dispatch('toggleLogin', {status: true});
+    return res;
+  }
+},
+
+err => {
+  const errRes = err.response;
+  console.log(`Error in interceptor`);
+  console.log(`Error: ${err}`);
+  if (errRes && errRes.staus == 401) {
+    window.localStorage.removeItem('fitness_token');
+  }
+  swal.fire({
+    type: 'error',
+    title: 'Oops...',
+    text: `Something wrong with server: \n${err}`,
+  });
+  router.push('/home');
+  return Promise.reject(err);
+})
+
+const oauthHttp = axios.create({
+  baseURL: 'http://oauth.scarlet-temp.tk/'
+})
+oauthHttp.interceptors.request.use(req => {
+  if ('oauth_token' in window.localStorage) {
+    console.log('Using oauth token')
+    const oauth_token = window.localStorage.getItem('oauth_token');
+    req.headers['Authorization'] = `Bearer ${oauth_token}`;
+  }
+  return req;
+})
+oauthHttp.interceptors.response.use(res => {
+  if (res.status == 200) {
+    // if token not exist, store the token
+    if (res.data.token) {
+      console.log(`Initilizing token: ${res.data.token}`);
       window.localStorage.setItem('fitness_token', res.data.token);
       const Toast = swal.mixin({
         toast: true,
@@ -48,30 +101,18 @@ fitnessHttp.interceptors.response.use(res => {
     }
     // setup login status
     store.dispatch('toggleLogin', {status: true});
+    store.dispatch('setOauthLogin', true);
     return res;
   }
-},
-err => {
-  const errRes = err.response;
-  console.log(`Error in interceptor`);
-  console.log(`Error: ${err}`);
-  if (errRes && errRes.staus == 401) {
-    window.localStorage.removeItem('fitness_token');
-  }
-  swal.fire({
-    type: 'error',
-    title: 'Oops...',
-    text: `Something wrong with server: \n${err}`,
-  });
-  router.push('/home');
-  return Promise.reject(err);
 })
+
 Vue.config.productionTip = false
 // register
 Vue.use(GlobalComponents);
 Vue.use(MaterialDashboard);
 Vue.prototype.$axios = axios;
 Vue.prototype.$fitnessHttp = fitnessHttp;
+Vue.prototype.$oauthHttp = oauthHttp;
 Vue.prototype.$Chartist = Chartist;
 
 new Vue({
